Refactor scores and notebooks
